+++ crates/routevn-exporter-jni/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "routevn-exporter-jni"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+name = "routevn_exporter_jni"
+crate-type = ["cdylib"]
+
+[dependencies]
+jni = "0.21"
+routevn-exporter = { path = "../routevn-exporter" }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
